Upgradeables file now reading from top-bottom

+++ betteraptupgrade/betteraptupgrade.sh
@@ -8,20 +8,23 @@ listlength=$(($listlength - 0))
 
 # Read input of amount to be upgraded
 echo $listlength available updates
-echo How many to upgrade:
-read amount
+read -p "How many to upgrade: " amount
+
+read -p "Clean between each upgrade? Y/n[Y] " clean
+routers=${clean:-Y}
 
 # Get file line length
 excludelength=$(wc -l < src/excludes.txt)
+row=1
 
 # Break when end of file reached
-while [ $listlength -gt 1 ] && [ $amount -gt 0 ]
+while [ $listlength -gt $row ] && [ $amount -gt 0 ]
 do
   # Revalue excluded variable to false
   excluded="false"
 
   # Read current file row
-  package=$(awk -v linenum=$listlength 'NR == linenum {print; exit}' src/upgradeables.txt)
+  package=$(awk -v linenum=$row 'NR == linenum {print; exit}' src/upgradeables.txt)
   package="$(echo "$package" | cut -d'/' -f1)"
   excludelengthtemp=$(($excludelength - 0))
 
@@ -39,25 +42,28 @@ do
     then
       # If excluded=true then apt install package won't execute
       excluded="true"
-      echo $package excluded
+      echo -e "\n$package excluded\n"
       break;
     fi
     excludelengthtemp=$(($excludelengthtemp - 1))
   done
 
-  # Subtract listlength to change file line position
+  # Move to next line (row + 1)
   # Subtract amount to break upgrade loop when wished amount of packages have been upgraded
-  listlength=$(($listlength - 1))
+  row=$(($row + 1))
   amount=$(($amount - 1))
 
   if [ "$excluded" == "false" ]
   then
     apt-get install -y $package
   fi
-  
-  echo Cleaning
+
   # Clean cache between each upgrade
-  apt-get -y clean
-  apt-get -y autoclean
-  apt-get -y autoremove
+  if [ "$clean" != "n" ]
+  then
+    echo Cleaning
+    apt-get -y clean
+    apt-get -y autoclean
+    apt-get -y autoremove
+  fi
 done
